Guess what? More refactoring.
